fix(cp): return fp32 from the empty response chunk to keep the CP all-reduce uniform
calculate_log_probs_and_entropy returns fp32 from every branch that computes something,
because compute_log_probs and compute_entropy both upcast. The branch for "this rank's
slice contains no response token" returned logits.dtype instead.

That is invisible while logits are fp32. Under context parallelism with a long prompt and a
short response, some CP ranks hold no response token and others do, and their results meet
in one all-reduce over the concatenated response tensors. With bf16 logits the two groups
disagree on byte count and the collective never completes -- every rank sits in ALLREDUCE
until the watchdog kills the job at 600 s, with nothing in the log pointing at dtype.

Backport of the fix that landed upstream in #2818, so this branch's pinned base carries it.

+++ miles/backends/training_utils/loss_hub/math_utils.py
@@ -984,9 +984,15 @@ def calculate_log_probs_and_entropy(
             if with_entropy:
                 entropy = compute_entropy(logits)
     else:
-        log_prob = logits.new_zeros((0,))
+        # compute_log_probs and compute_entropy both upcast their input, so every branch above
+        # returns fp32 whatever the logits dtype is. This one has to match: with bf16 logits it
+        # would otherwise return bf16, and under context parallelism a rank whose slice holds no
+        # response token takes this branch while the others do not. They then meet in one
+        # all-reduce over the concatenated response tensors, disagreeing on byte count. That is a
+        # hang, not an error -- the collective never completes and the job dies to the watchdog.
+        log_prob = logits.new_zeros((0,), dtype=torch.float32)
         if with_entropy:
-            entropy = logits.new_zeros((0,))
+            entropy = logits.new_zeros((0,), dtype=torch.float32)
 
     return log_prob, entropy
